+++ packages/cad-simple-viewer/src/command/AcApSysVarCmd.ts
@@ -1,0 +1,27 @@
+import { AcDbSysVarManager } from '@mlightcad/data-model'
+
+import { AcApContext, AcApDocManager } from '../app'
+import { AcEdCommand, AcEdPromptStringOptions } from '../editor'
+import { AcApI18n } from '../i18n'
+
+/**
+ * Command for modifying value of one system variable. All of system variables share
+ * this command.
+ */
+export class AcApSysVarCmd extends AcEdCommand {
+  /**
+   * Executes the command to modify the value of one system variable.
+   *
+   * @param context - The application context containing the view
+   */
+  async execute(_context: AcApContext) {
+    const prompt = new AcEdPromptStringOptions(AcApI18n.t('jig.sysvar.prompt'))
+    AcEdPromptStringOptions
+    const value = await AcApDocManager.instance.editor.getString(prompt)
+    const sysVarManager = AcDbSysVarManager.instance()
+    const sysVar = sysVarManager.getDescriptor(this.globalName)
+    if (sysVar) {
+      sysVarManager.setVar(this.globalName, value)
+    }
+  }
+}