+++ packages/cad-simple-viewer/src/i18n/en/jig.ts
@@ -2,5 +2,8 @@ export default {
   circle: {
     center: 'Specify the center of circle:',
     radius: 'Specify the radius of circle:'
+  },
+  sysvar: {
+    prompt: 'Please input new value'
   }
 }