+++ packages/cad-simple-viewer/src/i18n/zh/jig.ts
@@ -2,5 +2,8 @@ export default {
   circle: {
     center: '指定圆的圆心：',
     radius: '指定圆的半径：'
+  },
+  sysvar: {
+    prompt: '请输入新的值'
   }
 }