+++ packages/three-renderer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @mlightcad/three-renderer
 
+## 1.2.5
+
+### Patch Changes
+
+- feat: refine input mangager and jig system
+
 ## 1.2.4
 
 ### Patch Changes